all: use probe services URLs
This is again yak shaving when working on the testing described
at https://github.com/ooni/backend/issues/446.

I noticed we were still using old names and figured it would rather
be a better idea to instead just switch to probe services.

Of course, we're still using legacy APIs, but at least we're not
using anymore names that really belong to the past.